Altered migrations, created models (#9)
* removed previous migrations

* added institution table

* updated to 5.8 then downgraded back to 5.5

* created company table

* created recruiter table

* created job category table

* created job table

* Revert "created job table". I forgot to add is_deleted.

This reverts commit 7c908de9a23ee775c2df6b3af06833681a9cd3fa.

* created job table with is_deleted

* created jobseeker table

* created cv table

* created resume table

* created job experience table

* Revert "created job experience table". is_deleted is forgotten

This reverts commit f457e23547097314da3bd085130cc922eee5abe8.

* created job experience table with is_deleted

* Revert "created job table with is_deleted". Circular design encountered. Separating default documents and submitted documents entity"

This reverts commit fb2b6be466830d31624159b6212b8ed7e2f6ff04.

* added job_application, job_experience, and renamed job_experience

* merged jobseeker, recruiter, and user table

* altered database according to technical side

* created models

* added gitignore for root directory

+++ jobseekr/database/migrations/2019_10_17_181200_create_job_application_table.php
@@ -16,7 +16,7 @@ class CreateJobApplicationTable extends Migration
         Schema::create('job_application', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->bigInteger('jobseeker_id')->unsigned();
-            $table->integer('job_id')->unsigned();
+            $table->bigInteger('job_id')->unsigned();
             $table->string('summary', 200);
             $table->integer('cv_id')->unsigned();
             $table->integer('resume_id')->unsigned();
